implement all changes for the redesign
