+++ assets/svg/reset_password_icon.js
@@ -1,0 +1,22 @@
+import * as React from "react";
+import Svg, { G, Rect, Path, Defs } from "react-native-svg";
+const ResetPasswordIcon = (props) => (
+    <Svg
+        width={128}
+        height={128}
+        viewBox="0 0 128 128"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+        {...props}
+    >
+        <G filter="url(#filter0_d_1104_709)">
+            <Rect x={4} width={120} height={120} rx={20} fill="white" />
+        </G>
+        <Path
+            d="M69.4086 21.6324C47.9882 21.0432 30.4394 38.2132 30.4394 59.5074H22.9065C21.0128 59.5074 20.0869 61.7799 21.4336 63.0845L33.1749 74.8257C34.0165 75.6674 35.3211 75.6674 36.1628 74.8257L47.9461 63.0845C49.2086 61.7799 48.2828 59.5074 46.389 59.5074H38.8561C38.8561 43.137 52.3228 29.7965 68.7353 30.049C84.3482 30.2595 97.5203 43.4315 97.7728 59.0865C98.0253 75.5411 84.7269 88.9657 68.3144 88.9657C61.6232 88.9657 55.479 86.7353 50.5132 82.9478C48.8299 81.6853 46.4732 81.8115 45.0003 83.3265C43.1907 85.1361 43.359 88.124 45.379 89.639C51.7336 94.4786 59.6874 97.3824 68.3144 97.3824C89.6086 97.3824 106.779 79.8336 106.189 58.4553C105.642 38.6761 89.1457 22.1795 69.4086 21.6324ZM76.7311 55.299V51.0907C76.7311 46.4615 72.9436 42.674 68.3144 42.674C63.6853 42.674 59.8978 46.4615 59.8978 51.0907V55.299C57.5832 55.299 55.6894 57.1928 55.6894 59.5074V72.1324C55.6894 74.447 57.5832 76.3407 59.8978 76.3407H76.7311C79.0457 76.3407 80.9394 74.447 80.9394 72.1324V59.5074C80.9394 57.1928 79.0457 55.299 76.7311 55.299ZM72.5228 55.299H64.1061V51.0907C64.1061 48.7761 65.9999 46.8824 68.3144 46.8824C70.629 46.8824 72.5228 48.7761 72.5228 51.0907V55.299Z"
+            fill="#112D4E"
+        />
+        <Defs></Defs>
+    </Svg>
+);
+export default ResetPasswordIcon;
